Refactor: Re-Organization of SwerveDrive.cpp
I was never happy with how SwerveDrive.cpp was organized, it was over 200 lines
and most of those were for configuration of the motors
so I took the time to rewrite SwerveDrive.cpp how I kinda always wanted it to look.
To do this I made 2 helper classes: NeoSpark.cpp and NeoSet.cpp. NeoSpark handles the
configuration of the Neo and holds the objects for the motor and the relative encoder
(if enabled). NeoSet handles multiple NeoSpark sets.
All of this means that I was able to knock the number of object members in
SwerveDrive.h from 16 to 5. There also used to be about 230 lines in SwerveDrive.cpp and
now there are only 77.