Build successful
Needed to remove brake/coast mode for the Kraken drive motors from the wheel calibration procedure since the Kraken documentation said this should not be set every loop.  The Kraken drive motors are configured for brake mode (set in Constants file) when the motor outputs are neutral or disabled.

To-do:
* Review everything with Janelyn!
* Define the CAN update frequency for encorder feedback from the Constant file and pass to the Kraken function calls.
* Optimize CAN bus utilization
https://pro.docs.ctr-electronics.com/en/latest/docs/api-reference/api-usage/status-signals.html