Purpose of changes:
1) Update the maximum motor speed from 5500 (Neo) to 6200 (Kraken) RPM
2) Add multiply by 60 in KrakenTalon.cpp to convert from Rev/Sec to Rev/Min
3) Increase steering current limit from 20 to 60 Amps

After the Simulink build, the VS code build was failing because the Matlab script for SimulinkSmartDashboardInterface.cpp was including SimulinkSmartDashboardInterface.hh which included lib\include\Component.hh which did not exist.  Changed the Matlab script to use .h extensions instead of .hh extensions so it could find the files that exist.